Completely overhaul the web interface.
Sorry if you liked the LCARS theme.

The previous interface was extremely 'fragile' to any
modifications and additions that were to be made; the theme
was 'just enough' to display what was present and so
was difficult to update and have things fit in.

Also the JS code plumbing was (and still is) a bit of a mess.

The previous interface also didn't scale well across a
large number of devices, and in general, the LCARS interface
made it difficult to add content and still have enough
space to fit everything in. By contrast, using a standard
Bootstrap theme means that by design, it will scale well
across devices (fluid layout), and in general is more well
documented and easier to modify/update.

Squashed commits below:
Update Leaflet to git master (fixes IE zoom bug)